# shopping-services/src/main/resources/swagger/system-order-and-payments-0.1.yml
openapi: 3.0.1
info:
  title: System Order and Payments
  description: This is Documentation API
  version: "0.1"
paths:
...
  /product/addProduct:
    post:
      operationId: addProduct
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Product"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
  /product/delete/{idProduct}:
    delete:
      operationId: delete
      parameters:
      - name: idProduct
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
...
  /product/searchProduct/{idProduct}:
    get:
      operationId: search
      parameters:
      - name: idProduct
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
  /product/updateOrderInProduct/{idProduct}:
    patch:
      operationId: updateOrderInProduct
      parameters:
      - name: idProduct
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Product"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
...
components:
  schemas:
...
    Product:
      required:
      - idProduct
      - name
      - price
      - quantity
      type: object
      properties:
        idProduct:
          type: string
        name:
          type: string
        price:
          type: number
          format: double
        quantity:
          type: integer
          format: int32
    ProductResponse:
      required:
        - id
        - idProduct
        - name
        - price
        - quantity
      type: object
      properties:
        id:
          type: number
        idProduct:
          type: string
        name:
          type: string
        price:
          type: number
          format: double
        quantity:
          type: integer
          format: int32
...